# API_HOTEL_LARAVEL/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Hotel",
        "description": "API para gerenciar as atividades de hoteis",
        "contact": {
            "email": "italojefer55@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/"
        }
    ],
    "paths": {
...
        "/api/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Lista todos os pagamentos",
                "description": "Acesso permitido para administradores e recepcionistas",
                "operationId": "5d60a9222dff76d7cf262181c402186c",
                "responses": {
                    "200": {
                        "description": "Lista de pagamentos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Cadastra um novo pagamento",
                "description": "Acesso permitido para administradores e recepcionistas",
                "operationId": "d36e2be5fe281b6d1586dd9ce1306ab5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reserveId",
                                    "value",
                                    "method",
                                    "paid"
                                ],
                                "properties": {
                                    "reserveId": {
                                        "type": "string",
                                        "example": "R123"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 200
                                    },
                                    "method": {
                                        "type": "string",
                                        "example": "Credit Card"
                                    },
                                    "paid": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pagamento criado com sucesso"
                    },
                    "500": {
                        "description": "Falha ao criar pagamento"
                    }
                }
            }
        },
        "/api/payments/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Exibe os detalhes de um pagamento específico",
                "description": "Acesso permitido para administradores e recepcionistas",
                "operationId": "c3632c1a7d5e6eebb409bdeb974af413",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do pagamento"
                    },
                    "404": {
                        "description": "Falha ao buscar pagamento"
                    }
                }
            },
            "put": {
                "tags": [
                    "Payments"
                ],
                "summary": "Atualiza um pagamento",
                "description": "Acesso permitido para administradores e recepcionistas",
                "operationId": "c875f05cad3e5fce8d2a704a64e641db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reserveId",
                                    "value",
                                    "method",
                                    "paid"
                                ],
                                "properties": {
                                    "reserveId": {
                                        "type": "string",
                                        "example": "R123"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 200
                                    },
                                    "method": {
                                        "type": "string",
                                        "example": "Credit Card"
                                    },
                                    "paid": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pagamento atualizado com sucesso"
                    },
                    "500": {
                        "description": "Falha ao atualizar pagamento"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Payments"
                ],
                "summary": "Exclui um pagamento",
                "description": "Acesso permitido para administradores e recepcionistas",
                "operationId": "0a18c3c7854ce5ee449734bd2a0d7f9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagamento excluído com sucesso"
                    },
                    "500": {
                        "description": "Falha ao excluir pagamento"
                    }
                }
            }
        },
...
    },
...
}